packaged all into a crate and divided it into features
