+++ crates/moss-fs/Cargo.toml
@@ -2,6 +2,7 @@
 name = "moss_fs"
 version = "0.1.0"
 edition = "2021"
+publish = false
 
 [dependencies]
 dunce.workspace = true
