+++ crates/moss-html/Cargo.toml
@@ -2,6 +2,7 @@
 name = "moss_html"
 version = "0.1.0"
 edition = "2021"
+publish = false
 
 [lib]
 crate-type = ["cdylib", "rlib"]
