+++ crates/moss-themeconv/Cargo.toml
@@ -1,7 +1,8 @@
 [package]
-name = "workbench_service_theme"
+name = "moss_themeconv"
 version = "0.1.0"
 edition = "2021"
+publish = false
 
 [dependencies]
 desktop_models.workspace = true
@@ -9,3 +10,7 @@ serde = { workspace = true, features = ["derive"] }
 serde_json.workspace = true
 anyhow.workspace = true
 indexmap = { workspace = true, features = ["serde"] }
+
+[features]
+default = ["json"]
+json = []
