+++ crates/moss-themeconv/src/model.rs
@@ -1,5 +1,3 @@
-use std::collections::HashMap;
-
 use desktop_models::appearance::theming::ThemeType;
 use indexmap::IndexMap;
 use serde::{Deserialize, Serialize};
