+++ crates/moss-uikit/Cargo.toml
@@ -2,6 +2,7 @@
 name = "moss_uikit"
 version = "0.1.0"
 edition = "2021"
+publish = false
 
 [lib]
 crate-type = ["cdylib", "rlib"]
