--- internal/workbench/service/theme/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod theme_convert;
-pub mod theme_model;
